patchwise_parallel_512x512x604_multi_gpu/requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0
numpy<2.0
nibabel>=5.0.0
scipy>=1.10.0
omegaconf>=2.3.0
hydra-core>=1.3.0
einops>=0.6.0
tensorboard>=2.12.0
tqdm>=4.65.0
scikit-image>=0.20.0
